Backend/requirements.txt
fastapi
uvicorn
langchain
transformers
torch
accelerate
sentence-transformers
faiss-cpu
huggingface_hub
python-dotenv
psycopg2-binary
pydantic
bitsandbytes
